requirements.txt
certifi==2018.4.16
chardet==3.0.4
decorator==4.3.0
idna==2.7
ipython-genutils==0.2.0
jsonschema==2.6.0
jupyter-core==4.4.0
nbformat==4.4.0
numpy==1.15.0
oauthlib==2.1.0
pandas==0.23.3
plotly==3.1.0
PySocks==1.6.8
python-dateutil==2.7.3
pytz==2018.5
PyYAML==3.13
requests==2.19.1
requests-oauthlib==1.0.0
retrying==1.3.3
six==1.11.0
traitlets==4.3.2
tweepy==3.6.0
urllib3==1.23
